Split off the HTML decoder functionality into an extension.
