+++ RedditRealTime/RedditRealTime/View Layer/View Controllers/RisingStoriesViewController.swift
@@ -78,7 +78,7 @@ class RisingStoriesViewController: UIViewController, UITableViewDataSource, UITa
         
         guard let cell = tableView.dequeueReusableCell(withIdentifier: "PostTableViewCell") as? PostTableViewCell else { return UITableViewCell() }
         
-        cell.title.text = removeHTMLEntities(str: redditPostFetcher.posts[indexPath.row].title)?.string
+        cell.title.text = redditPostFetcher.posts[indexPath.row].title.removeHTMLEntities()?.string
         cell.upvotes.text = " \(redditPostFetcher.posts[indexPath.row].upvotes) Upvotes"
         cell.comments.text = " \(redditPostFetcher.posts[indexPath.row].commentCount) Comments"
         
@@ -88,25 +88,6 @@ class RisingStoriesViewController: UIViewController, UITableViewDataSource, UITa
     
     @IBAction func unwindToRisingStories(segue: UIStoryboardSegue) {
         
-    }
-    
-    private func removeHTMLEntities(str: String) -> NSAttributedString? {
-        
-        let data = str.data(using: .utf8)
-        
-        let options : [NSAttributedString.DocumentReadingOptionKey : Any] = [
-            .documentType: NSAttributedString.DocumentType.html,
-            .characterEncoding: String.Encoding.utf8.rawValue
-        ]
-        
-        do {
-            let attributedString = try NSAttributedString(data: data!, options: options, documentAttributes: nil)
-            return attributedString
-        } catch {
-            print(error)
-        }
-        
-        return nil
     }
